+++ BuildLogic/src/main/kotlin/build-logic.java-common-conventions.gradle.kts
@@ -12,8 +12,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-import org.swift.swiftjava.gradle.JavaLibraryPathUtils.javaLibraryPaths
+import java.util.*
 import java.io.*
+import kotlin.system.exitProcess
+import kotlinx.serialization.json.*
 
 plugins {
     java
@@ -41,6 +43,62 @@ testing {
 tasks.withType(JavaCompile::class).forEach {
     it.options.compilerArgs.add("--enable-preview")
     it.options.compilerArgs.add("-Xlint:preview")
+}
+
+
+fun getSwiftRuntimeLibraryPaths(): List<String> {
+    val process = ProcessBuilder("swiftc", "-print-target-info")
+        .redirectError(ProcessBuilder.Redirect.INHERIT)
+        .start()
+
+    val output = process.inputStream.bufferedReader().use { it.readText() }
+    val exitCode = process.waitFor()
+    if (exitCode != 0) {
+        System.err.println("Error executing swiftc -print-target-info")
+        exitProcess(exitCode)
+    }
+
+    val json = Json.parseToJsonElement(output)
+    val runtimeLibraryPaths = json.jsonObject["paths"]?.jsonObject?.get("runtimeLibraryPaths")?.jsonArray
+    return runtimeLibraryPaths?.map { it.jsonPrimitive.content } ?: emptyList()
+}
+
+/**
+ * Find library paths for 'java.library.path' when running or testing projects inside this build.
+ */
+// TODO: can't figure out how to share this code between BuildLogic/ and buildSrc/
+fun javaLibraryPaths(rootDir: File): List<String> {
+    val osName = System.getProperty("os.name").lowercase(Locale.getDefault())
+    val osArch = System.getProperty("os.arch")
+    val isLinux = osName.contains("linux")
+    val base = rootDir.path.let { "$it/" }
+
+    val projectBuildOutputPath =
+        if (isLinux) {
+            if (osArch == "amd64" || osArch == "x86_64")
+                "$base.build/x86_64-unknown-linux-gnu"
+            else
+                "$base.build/${osArch}-unknown-linux-gnu"
+        } else {
+            if (osArch == "aarch64")
+                "$base.build/arm64-apple-macosx"
+            else
+                "$base.build/${osArch}-apple-macosx"
+        }
+    val parentParentBuildOutputPath =
+        "../../$projectBuildOutputPath"
+
+
+    val swiftBuildOutputPaths = listOf(
+        projectBuildOutputPath,
+        parentParentBuildOutputPath
+    )
+
+    val debugBuildOutputPaths = swiftBuildOutputPaths.map { "$it/debug" }
+    val releaseBuildOutputPaths = swiftBuildOutputPaths.map { "$it/release" }
+    val swiftRuntimePaths = getSwiftRuntimeLibraryPaths()
+
+    return debugBuildOutputPaths + releaseBuildOutputPaths + swiftRuntimePaths
 }
 
 // Configure paths for native (Swift) libraries
