--- BuildLogic/src/main/kotlin/org/swift/swiftjava/gradle/DO_NOT_EDIT.txt
@@ -1,6 +1,0 @@
-DO NOT EDIT THIS PACKAGE
-
-It is copied from buildSrc/ during build.
-
-We cannot easily share the sources between BuildLogic and buildSrc so we automated the copy.
-If this could be shared in some nicer way, please improve the build to do so :-)
